requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
jinja2>=3.1.4
python-dotenv>=1.0.1
openai>=1.51.0
python-multipart>=0.0.12
pydantic>=2.10.0
httpx>=0.27.2
duckduckgo-search>=6.2.0
